Clean some old stuff out
