--- web/templates/login.tmpl.html
@@ -1,82 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    <title>WebAuthn Login</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            max-width: 600px;
-            margin: 50px auto;
-            padding: 20px;
-        }
-        .form-group {
-            margin-bottom: 20px;
-        }
-        button {
-            background-color: #007bff;
-            color: white;
-            padding: 12px 24px;
-            border: none;
-            border-radius: 4px;
-            cursor: pointer;
-            font-size: 16px;
-        }
-        button:hover {
-            background-color: #0056b3;
-        }
-        button:disabled {
-            background-color: #6c757d;
-            cursor: not-allowed;
-        }
-        .error-message {
-            color: #dc3545;
-            background-color: #f8d7da;
-            border: 1px solid #f5c6cb;
-            padding: 10px;
-            border-radius: 4px;
-            margin-bottom: 20px;
-            display: none;
-        }
-        .loading {
-            display: none;
-            color: #6c757d;
-            font-style: italic;
-        }
-        .login-container {
-            text-align: center;
-            padding: 40px 20px;
-        }
-        .login-description {
-            color: #6c757d;
-            margin-bottom: 30px;
-        }
-    </style>
-</head>
-
-<body>
-    <!-- Session ID for the JavaScript -->
-    <div data-session-id="{{ .SessionID }}"></div>
-
-    <div class="login-container">
-        <h1>WebAuthn Login</h1>
-
-        <p class="login-description">
-            Use your security key to sign in. No username or password required.
-        </p>
-
-        <div id="error-message" class="error-message"></div>
-
-        <form id="login-form">
-            <div class="form-group">
-                <button type="submit" id="login-button">Sign in with Security Key</button>
-                <span id="login-loading" class="loading">Processing...</span>
-            </div>
-        </form>
-    </div>
-
-    <script src="/public/webauthn.js"></script>
-</body>
-
-</html>
